+++ src/com/alibaba/craftsman/metrics/Measurable.java
@@ -7,5 +7,5 @@ package com.alibaba.craftsman.metrics;
  * @date 2018-07-04 1:32 PM
  */
 public interface Measurable {
-    public long calculateScore();
+    public double calculateScore();
 }
